Use `pytest_sessionfinish` instead of `atexit`
With 3.12, it is no longer possible to create a new thread in an
`atexit` handler. Stopping the monitor process requires writing a
message in the queue, which itself may require the creation of a thread.

This commit will remove the need to have an `atexit` handler, by using
the `pytest_sessionfinish` hook instead, which should work in our case.

Signed-off-by: Gus Monod <gmonod1@bloomberg.net>